feat. linkContainer 분리, 검색 결과 없을경우 재검색 처리
