+++ src/containers/LinkContainer/index.ts
@@ -1,0 +1,3 @@
+import LinkContainer from './LinkContainer'
+
+export default LinkContainer
